Description: fix use of portals/lnet pid
Details    : incorrect use of portals/lnet pid caused them to get out of
	     sync and would result in silent dropping of RPC messages
b=10074
r=eeb (original patch)

Severity   : major
Frequency  : Infiniband IB LND only
Description: iiblnd wasn't mapping all memory
Details    : iiblnd wasn't mapping all memory, resulting in comms errors
	     on some architectures/memory configs
b=9776
r=eeb (original patch)

+++ lnet/include/libcfs/portals_lib.h
@@ -29,6 +29,8 @@
 #include <libcfs/linux/portals_lib.h>
 #elif defined(__APPLE__)
 #include <libcfs/darwin/portals_lib.h>
+#elif defined(__WINNT__)
+#include <libcfs/winnt/portals_lib.h>
 #else
 #error Unsupported Operating System
 #endif
@@ -68,7 +70,7 @@ static inline int size_round0(int val)
 
 static inline size_t round_strlen(char *fset)
 {
-        return size_round(strlen(fset) + 1);
+        return (size_t)size_round((int)strlen(fset) + 1);
 }
 
 #define LOGL(var,len,ptr)                                       \
